video 1, 2 and 3
(i had to do this because it wouldn't let me commit due to it thinking i didnt update anything.) 1st video introduces us to module. on how to use them in angular and what its used for. 2nd video is a refresher for ngFor and ngif. all which can be added towards the app component. and the 3rd video was on ngClass and ngStyle recap. explaining simalar to the last video. but is more on how to get css information into the component